refactor(VariantSelector): split test file into mode-specific files
